requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
unstructured[all-docs]
pillow
lxml
python-dotenv
rich
tqdm
